requirements.txt
alembic==1.14.0
aniso8601==9.0.1
argon2-cffi 
argon2-cffi-bindings 
asyncio==3.4.3
attrs==24.2.0
bcrypt 
binho-host-adapter==0.1.6
certifi==2024.8.30
cffi==1.17.1
cycler==0.12.1
dnspython==2.7.0
et_xmlfile==2.0.0
Flask 
flask-cors 
Flask-JWT-Extended
Flask-Mail 
flatbuffers==24.3.25
flexcache==0.3
flexparser==0.4
fonttools==4.55.0
greenlet==3.1.1
gunicorn==23.0.0
humanfriendly==10.0
idna==3.10
imageio==2.36.0
itsdangerous==2.2.0
joblib==1.4.2
jsonschema
jsonschema-specifications
kiwisolver==1.4.7
lazy_loader==0.4
llvmlite==0.43.0
Mako==1.3.6
MarkupSafe
matplotlib
MetPy==1.6.3
mpld3==0.5.10
mpmath==1.3.0
networkx
numba
numpy
onnxruntime
opencv-python
opencv-python-headless
openpyxl==3.1.5
packaging==24.2
paho-mqtt
pandas
pillow
Pint==0.24.4
platformdirs==4.3.6
pooch==1.8.2
protobuf==5.28.3
psutil==6.1.0
pycparser==2.22
pyftdi==0.56.0
PyJWT
PyMatting==1.1.13
pymongo
PyQRCode==1.2.1
python-dateutil==2.9.0.post0
python-dotenv 
pytz==2024.2
pyudev==0.24.3
pyzbar==0.1.9
referencing==0.35.1
rembg==2.0.59
requests==2.32.3
rpds-py==0.21.0
scikit-image==0.24.0
scikit-learn==1.5.2
scipy 
six 
sympy 
threadpoolctl 
tifffile 
tqdm 
traitlets 
typing_extensions 
tzdata 
urllib3 
Werkzeug 
xarray 
